# Optimize chunk prefill performance