Some offline benchmarking:
(1) vllm + chunk prefill: 6.2s
(2) vllm + chunk prefill + lmcache: 10.7s (1st request), 1.3s (2nd), 1.3s (3rd).

@Shaoting-Feng Please test the e2e results as well.